refactor: move background color fn
